+++ sdk/ai/azure-ai-projects/samples/agents/async_samples/sample_agents_code_interpreter_attachment_enterprise_search_async.py
@@ -3,8 +3,6 @@
 # Licensed under the MIT License.
 # ------------------------------------
 """
-FILE: sample_agents_code_interpreter_attachment_enterprise_search_async.py
-
 DESCRIPTION:
     This sample demonstrates how to use agent operations with code interpreter from
     the Azure Agents service using a synchronous client.
@@ -17,11 +15,10 @@ USAGE:
     pip install azure-ai-projects azure-identity aiohttp
 
     Set this environment variables with your own values:
-    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Studio Project.
+    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Foundry project.
 """
 import asyncio
 import os
-
 from azure.ai.projects.aio import AIProjectClient
 from azure.ai.projects.models import (
     CodeInterpreterTool,
@@ -30,11 +27,6 @@ from azure.ai.projects.models import (
     VectorStoreDataSourceAssetType,
 )
 from azure.identity.aio import DefaultAzureCredential
-
-
-# Create an Azure AI Client from a connection string, copied from your AI Studio project.
-# At the moment, it should be in the format "<HostName>;<AzureSubscriptionId>;<ResourceGroup>;<HubName>"
-# Customer needs to login to Azure subscription via Azure CLI and set the environment variables
 
 
 async def main():