+++ sdk/ai/azure-ai-projects/samples/agents/async_samples/sample_agents_stream_iteration_async.py
@@ -4,8 +4,6 @@
 # ------------------------------------
 
 """
-FILE: sample_agents_stream_iteration_async.py
-
 DESCRIPTION:
     This sample demonstrates how to use agent operations with interation in streaming from
     the Azure Agents service using a asynchronous client.
@@ -18,22 +16,19 @@ USAGE:
     pip install azure-ai-projects azure-identity aiohttp
 
     Set this environment variables with your own values:
-    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Studio Project.
+    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Foundry project.
 """
 import asyncio
 
 from azure.ai.projects.aio import AIProjectClient
 from azure.ai.projects.models import AgentStreamEvent
-from azure.ai.projects.models import MessageDeltaChunk, MessageDeltaTextContent, RunStep, ThreadMessage, ThreadRun
+from azure.ai.projects.models import MessageDeltaChunk, RunStep, ThreadMessage, ThreadRun
 from azure.identity.aio import DefaultAzureCredential
 
 import os
 
 
 async def main() -> None:
-    # Create an Azure AI Client from a connection string, copied from your AI Studio project.
-    # At the moment, it should be in the format "<HostName>;<AzureSubscriptionId>;<ResourceGroup>;<HubName>"
-    # Customer needs to login to Azure subscription via Azure CLI and set the environment variables
     async with DefaultAzureCredential() as creds:
         async with AIProjectClient.from_connection_string(
             credential=creds, conn_str=os.environ["PROJECT_CONNECTION_STRING"]
@@ -52,20 +47,16 @@ async def main() -> None:
             print(f"Created message, message ID {message.id}")
 
             async with await project_client.agents.create_stream(thread_id=thread.id, assistant_id=agent.id) as stream:
-                async for event_type, event_data in stream:
+                async for event_type, event_data, _ in stream:
 
                     if isinstance(event_data, MessageDeltaChunk):
-                        for content_part in event_data.delta.content:
-                            if isinstance(content_part, MessageDeltaTextContent):
-                                text_value = content_part.text.value if content_part.text else "No text"
-                                print(f"Text delta received: {text_value}")
+                        print(f"Text delta received: {event_data.text}")
 
                     elif isinstance(event_data, ThreadMessage):
                         print(f"ThreadMessage created. ID: {event_data.id}, Status: {event_data.status}")
 
                     elif isinstance(event_data, ThreadRun):
                         print(f"ThreadRun status: {event_data.status}")
-
                     elif isinstance(event_data, RunStep):
                         print(f"RunStep type: {event_data.type}, Status: {event_data.status}")