+++ sdk/ai/azure-ai-projects/samples/agents/sample_agents_functions_with_azure_monitor_tracing.py
@@ -4,12 +4,10 @@
 # ------------------------------------
 
 """
-FILE: sample_agents_functions_with_azure_monitor_tracing.py
-
 DESCRIPTION:
     This sample demonstrates how to use basic agent operations with function tools from
     the Azure Agents service using a synchronous client with Azure Monitor tracing.
-    View the results in the "Tracing" tab in your Azure AI Studio project page.
+    View the results in the "Tracing" tab in your Azure AI Foundry project page.
 
 USAGE:
     python sample_agents_functions_with_azure_monitor_tracing.py
@@ -19,7 +17,7 @@ USAGE:
     pip install azure-ai-projects azure-identity opentelemetry-sdk azure-monitor-opentelemetry
 
     Set these environment variables with your own values:
-    * PROJECT_CONNECTION_STRING - The Azure AI Project connection string, as found in your AI Studio Project.
+    * PROJECT_CONNECTION_STRING - The Azure AI Project connection string, as found in your AI Foundry project.
     * AZURE_TRACING_GEN_AI_CONTENT_RECORDING_ENABLED - Optional. Set to `true` to trace the content of chat
       messages, which may contain personal data. False by default.
 """
@@ -32,11 +30,6 @@ from azure.ai.projects.models import FunctionTool, RequiredFunctionToolCall, Sub
 from opentelemetry import trace
 from azure.monitor.opentelemetry import configure_azure_monitor
 
-
-# Create an AI Project Client from a connection string, copied from your AI Studio project.
-# At the moment, it should be in the format "<HostName>;<AzureSubscriptionId>;<ResourceGroup>;<HubName>"
-# Customer needs to login to Azure subscription via Azure CLI and set the environment variables
-
 project_client = AIProjectClient.from_connection_string(
     credential=DefaultAzureCredential(), conn_str=os.environ["PROJECT_CONNECTION_STRING"]
 )
@@ -45,7 +38,7 @@ project_client = AIProjectClient.from_connection_string(
 application_insights_connection_string = project_client.telemetry.get_connection_string()
 if not application_insights_connection_string:
     print("Application Insights was not enabled for this project.")
-    print("Enable it via the 'Tracing' tab in your AI Studio project page.")
+    print("Enable it via the 'Tracing' tab in your AI Foundry project page.")
     exit()
 configure_azure_monitor(connection_string=application_insights_connection_string)