+++ sdk/ai/azure-ai-projects/samples/agents/sample_agents_openapi.py
@@ -4,8 +4,6 @@
 # ------------------------------------
 
 """
-FILE: sample_agents_openapi.py
-
 DESCRIPTION:
     This sample demonstrates how to use agent operations with the 
     OpenAPI tool from the Azure Agents service using a synchronous client.
@@ -19,7 +17,7 @@ USAGE:
     pip install azure-ai-projects azure-identity jsonref
 
     Set this environment variables with your own values:
-    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Studio Project.
+    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Foundry project.
 """
 
 import os
@@ -28,32 +26,26 @@ from azure.ai.projects import AIProjectClient
 from azure.identity import DefaultAzureCredential
 from azure.ai.projects.models import OpenApiTool, OpenApiAnonymousAuthDetails
 
-
-# Create an Azure AI Client from a connection string, copied from your AI Studio project.
-# At the moment, it should be in the format "<HostName>;<AzureSubscriptionId>;<ResourceGroup>;<HubName>"
-# Customer needs to login to Azure subscription via Azure CLI and set the environment variables
-
 project_client = AIProjectClient.from_connection_string(
     credential=DefaultAzureCredential(),
     conn_str=os.environ["PROJECT_CONNECTION_STRING"],
 )
 
-with open('./weather_openapi.json', 'r') as f:
+with open("./weather_openapi.json", "r") as f:
     openapi_spec = jsonref.loads(f.read())
 
 # Create Auth object for the OpenApiTool (note that connection or managed identity auth setup requires additional setup in Azure)
 auth = OpenApiAnonymousAuthDetails()
 
 # Initialize agent OpenApi tool using the read in OpenAPI spec
-openapi = OpenApiTool(name="get_weather", spec=openapi_spec, description="Retrieve weather information for a location", auth=auth)
+openapi = OpenApiTool(
+    name="get_weather", spec=openapi_spec, description="Retrieve weather information for a location", auth=auth
+)
 
 # Create agent with OpenApi tool and process assistant run
 with project_client:
     agent = project_client.agents.create_agent(
-        model="gpt-4o-mini",
-        name="my-assistant",
-        instructions="You are a helpful assistant",
-        tools=openapi.definitions
+        model="gpt-4o-mini", name="my-assistant", instructions="You are a helpful assistant", tools=openapi.definitions
     )
     print(f"Created agent, ID: {agent.id}")