+++ sdk/ai/azure-ai-projects/samples/agents/sample_agents_stream_eventhandler.py
@@ -4,8 +4,6 @@
 # ------------------------------------
 
 """
-FILE: sample_agents_stream_eventhandler.py
-
 DESCRIPTION:
     This sample demonstrates how to use agent operations with an event handler in streaming from
     the Azure Agents service using a synchronous client.
@@ -18,7 +16,7 @@ USAGE:
     pip install azure-ai-projects azure-identity
 
     Set this environment variables with your own values:
-    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Studio Project.
+    PROJECT_CONNECTION_STRING - the Azure AI Project connection string, as found in your AI Foundry project.
 """
 
 import os
@@ -27,7 +25,6 @@ from azure.identity import DefaultAzureCredential
 
 from azure.ai.projects.models import (
     AgentEventHandler,
-    MessageDeltaTextContent,
     MessageDeltaChunk,
     ThreadMessage,
     ThreadRun,
@@ -36,11 +33,6 @@ from azure.ai.projects.models import (
 
 from typing import Any
 
-
-# Create an Azure AI Client from a connection string, copied from your AI Studio project.
-# At the moment, it should be in the format "<HostName>;<AzureSubscriptionId>;<ResourceGroup>;<HubName>"
-# Customer needs to login to Azure subscription via Azure CLI and set the environment variables
-
 project_client = AIProjectClient.from_connection_string(
     credential=DefaultAzureCredential(),
     conn_str=os.environ["PROJECT_CONNECTION_STRING"],
@@ -48,30 +40,28 @@ project_client = AIProjectClient.from_connection_string(
 
 
 # [START stream_event_handler]
-class MyEventHandler(AgentEventHandler):
+class MyEventHandler(AgentEventHandler[str]):
+
     def on_message_delta(self, delta: "MessageDeltaChunk") -> None:
-        for content_part in delta.delta.content:
-            if isinstance(content_part, MessageDeltaTextContent):
-                text_value = content_part.text.value if content_part.text else "No text"
-                print(f"Text delta received: {text_value}")
+        self.custom_data = f"Text delta received: {delta.text}"
 
     def on_thread_message(self, message: "ThreadMessage") -> None:
-        print(f"ThreadMessage created. ID: {message.id}, Status: {message.status}")
+        self.custom_data = f"ThreadMessage created. ID: {message.id}, Status: {message.status}"
 
     def on_thread_run(self, run: "ThreadRun") -> None:
-        print(f"ThreadRun status: {run.status}")
+        self.custom_data = f"ThreadRun status: {run.status}"
 
     def on_run_step(self, step: "RunStep") -> None:
-        print(f"RunStep type: {step.type}, Status: {step.status}")
+        self.custom_data = f"RunStep type: {step.type}, Status: {step.status}"
 
     def on_error(self, data: str) -> None:
-        print(f"An error occurred. Data: {data}")
+        self.custom_data = f"An error occurred. Data: {data}"
 
     def on_done(self) -> None:
-        print("Stream completed.")
+        self.custom_data = "Stream completed."
 
     def on_unhandled_event(self, event_type: str, event_data: Any) -> None:
-        print(f"Unhandled Event Type: {event_type}, Data: {event_data}")
+        self.custom_data = f"Unhandled Event Type: {event_type}, Data: {event_data}"
 
 
 # [END stream_event_handler]
@@ -94,7 +84,8 @@ with project_client:
     with project_client.agents.create_stream(
         thread_id=thread.id, assistant_id=agent.id, event_handler=MyEventHandler()
     ) as stream:
-        stream.until_done()
+        for _, _, custom_data in stream:
+            print(custom_data)
     # [END create_stream]
 
     project_client.agents.delete_agent(agent.id)