+++ sdk/ai/azure-ai-projects/CHANGELOG.md
@@ -1,14 +1,21 @@
 # Release History
 
-## 1.0.0b3 (Unreleased)
+## 1.0.0b3 (2024-12-13)
 
 ### Features Added
 
 * Added objects used to call Azure Functions.
-* Redesigned streaming and event handlers for agents 
+* Redesigned streaming and event handlers for agents.
 * Add `parallel_tool_calls` parameter to allow parallel tool execution.
+* Added `BingGroundingTool` for Agents to use against a Bing API Key connection.
+* Added `AzureAiSearchTool` for Agents to use against an Azure AI Search resource.
+* Added `OpenApiTool` for Agents, which creates and executes a REST function defined by an OpenAPI spec.
+* Added new helper properties in `OpenAIPageableListOfThreadMessage`, `MessageDeltaChunk`, and `ThreadMessage`.
+* Rename "AI Studio" to "AI Foundry" in package documents and samples, following recent rebranding.
 
-### Bugs Fixed
+### Breaking Changes
+
+* The method `.agents.get_messages` was removed. Please use `.agents.list_messages` instead.
 
 ## 1.0.0b2 (2024-12-03)